requirements.txt
streamlit
langchain
langchain-groq
langchain-community
python-dotenv
pypdf
chromadb
sentence-transformers

